% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{check_reverted}
\alias{check_reverted}
\title{Check Revert Probabilities (defunct)}
\usage{
check_reverted()
}
\description{
\code{check_reverted} previously identified
if an edit was considered likely, by
the ORES models, to be reverted. This model has now been
deprecated; users should instead rely on \code{\link{check_damaging}}
}
