% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query.R
\name{list_wikis}
\alias{list_wikis}
\title{List Supported Projects}
\usage{
list_wikis(...)
}
\arguments{
\item{...}{further arguments to pass to httr's GET.}
}
\description{
\code{\link{list_wikis}} lists Wikimedia
projects that support some or all of the ORES models.
}
