% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ores.R
\docType{package}
\name{ores}
\alias{ores}
\alias{ores-package}
\title{Connector to the Objective Revision Evaluation Service (ORES)}
\description{
This package provides a connector to ORES, an AI project
to provide edit scoring for content on Wikipedia and other Wikimedia projects.
This lets a researcher identify if edits are likely to be reverted,
damaging, or made in good faith.
}
